+++ ISWebPOS/Models/MstItemPackage.cs
@@ -13,7 +13,7 @@ namespace ISWebPOS.Models
         public Int32 ItemId { get; set; }
         public Int32 PackageItemId { get; set; }
         public Int32 UnitId { get; set; }
-        public Double Quantity { get; set; }
+        public Decimal Quantity { get; set; }
         public Boolean IsOptional { get; set; }
     }
 }
