+++ ISWebPOS/Models/MstItemPrice.cs
@@ -12,7 +12,7 @@ namespace ISWebPOS.Models
         public Int32 Id { get; set; }
         public Int32 ItemId { get; set; }
         public String PriceDescription { get; set; }
-        public Double Price { get; set; }
-        public Double TriggerQuantity { get; set; }
+        public Decimal Price { get; set; }
+        public Decimal TriggerQuantity { get; set; }
     }
 }
