+++ ISWebPOS/Models/MstTax.cs
@@ -12,7 +12,7 @@ namespace ISWebPOS.Models
         public Int32 Id { get; set; }
         public String Code { get; set; }
         public String Tax { get; set; }
-        public Double Rate { get; set; }
+        public Decimal Rate { get; set; }
         public Int32 AccountId { get; set; }
 
     }
